fix: resolve sonar findings of the overtime statistics branch
The demo record schedule derived the accrued and reduced hours with int
arithmetic and handed the result to a long parameter, so an overflow
would have wrapped before the widening. personOffset and accrualHours
are longs now.

Besides that only readability findings: java.time.Month constants
instead of int literals, isNotEmpty() in front of the allSatisfy and
doesNotContain assertions so they cannot pass on an empty list, the two
anySatisfy assertions of the accrual/reduction test joined into one
chain, an unnamed catch parameter in toYear and an unused import in the
security test.

closes #6397

+++ src/main/java/org/synyx/urlaubsverwaltung/dev/OvertimeDemoRecords.java
@@ -57,7 +57,7 @@ final class OvertimeDemoRecords {
 
             final int monthIndex = month.getMonthValue() - 1;
             // spreads the persons apart so the company sums are not just one curve multiplied
-            final int personOffset = (int) (personId % 4);
+            final long personOffset = personId % 4;
             // the previous year is a bit weaker, otherwise both year curves would lie on top of each other
             final int yearOffset = month.getYear() < today.getYear() ? -1 : 0;
 
@@ -65,7 +65,7 @@ final class OvertimeDemoRecords {
             final LocalDate accrualStart = month.atDay(ACCRUAL_WEEK_START);
             if (!accrualStart.isAfter(today)) {
                 final LocalDate accrualEnd = earlier(month.atDay(ACCRUAL_WEEK_END), today);
-                final int accrualHours = Math.max(1, ACCRUAL_HOURS_BY_MONTH[monthIndex] + personOffset + yearOffset);
+                final long accrualHours = Math.max(1L, ACCRUAL_HOURS_BY_MONTH[monthIndex] + personOffset + yearOffset);
                 entries.add(new Entry(accrualStart, accrualEnd, hours(accrualHours)));
             }
